Make size arg of r_file_slurp() size_t (#16221)

+++ libr/main/r2agent.c
@@ -84,7 +84,7 @@ R_API int r_main_r2agent(int argc, char **argv) {
 			return usage (0);
 		}
 
-		int sz;
+		size_t sz;
 		pfile = r_file_slurp (httpauthfile, &sz);
 		if (pfile) {
 			so.authtokens = r_str_split_list (pfile, "\n", 0);
